Fix plugin rename, test w relative path
